+++ src/pages/posts/[slug].tsx
@@ -6,7 +6,7 @@ import { getPostBySlug, getPostSlugs } from "@/lib/posts";
 import type { Post } from "@/types/post";
 import { PostBody } from "../../components/PostBody";
 import { format } from "date-fns";
-
+import styled from "@emotion/styled";
 export async function getStaticPaths() {
   const slugs = getPostSlugs();
   return {
@@ -26,6 +26,20 @@ export async function getStaticProps({ params }: { params: { slug: string } }) {
   };
 }
 
+const Wrapper = styled(Flex)`
+  max-width: 800px;
+  flex: 1 1 800px;
+
+  padding: 24px;
+  margin: 24px 12px;
+  max-height: fit-content;
+
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 6px -1px,
+    rgba(0, 0, 0, 0.06) 0px 2px 4px -1px;
+  border-radius: 12px;
+  background-color: white;
+`;
+
 interface PostDetailPageProps {
   post: Post;
   content: string;
@@ -33,11 +47,11 @@ interface PostDetailPageProps {
 
 export default function PostDetailPage({ post, content }: PostDetailPageProps) {
   return (
-    <Flex direction="column" gap={20}>
-      <Typography as="h1" variant="title1">
+    <Wrapper direction="column" gap={20}>
+      <Typography as="h1" variant="title3">
         {post.title}
       </Typography>
-      <Typography as="p" variant="body6">
+      <Typography as="p" variant="body3">
         {format(post.date, "yyyy년 MM월 dd일")}
       </Typography>
       <Flex gap={6}>
@@ -46,6 +60,6 @@ export default function PostDetailPage({ post, content }: PostDetailPageProps) {
         ))}
       </Flex>
       <PostBody content={content} />
-    </Flex>
+    </Wrapper>
   );
 }
